Service2/requirements.txt
alembic==0.9.9
Flask_WTF==0.14.2
Flask_Cors==3.0.3
Flask_SQLAlchemy==2.3.2
Flask==0.12.2
Flask_RESTful==0.3.6
passlib==1.7.1
Flask_Mail==0.9.1
Flask_Migrate==2.1.1
Flask_Script==2.0.6
Flask_Security==3.0.0
itsdangerous==0.24
SQLAlchemy==1.2.6
Flask_Login==0.4.1
python-dotenv==0.8.2
PyJWT==1.6.1
